track(Base): allow deferring the processing of modified files. Fixes #168.
This permits the caller to decide what to do about files that appear to be
modified (based on the `mtime` criterion for Base, or on genuine modifications
for Core.Compiler and the stdlibs). One option would be to put their
processing inside an `@async`, or just discard them altogether.
Rebugger will likely take the latter option, forcing the user to explicitly
say `Revise.track(Base)` if s/he wants to process revisions.