Switch to relative paths and indexing by PkgId
This brings Revise into better concordance with Pkg and the way code-loading
works in Base. By using relative paths, we pave the way for relocating the
collection of files, for example when switching environments or between `free`
and `dev` modes.

This is not breaking in terms of exported interface, but any packages
that use Revise's internals may need modification.