Remove pkgdata from revise_dir_queued (fixes #339) (#343)
Passing the pkgdata to revise_dir_queued means that if two packages
are watching the same directory and the wrong callback handles it,
then changes can get dropped. This switches to an implementation in
which we save the PkgId associated with each file in a watched
directory, and then process all the pkgdatas together when handling
the directory.