Single traversal implementation
This should be more efficient and less annoying, as each
deprecated binding gets accessed only once.

This also adds some refinements in the invalidation, and tests of functionality.
The only obvious weaknesses at this point are the bizarre need for printing
and the fact that some compiled instances don't have their inference results
cached.