Support user-specified callbacks
This feature is one half of what `Revise.entr` already does. The
function `entr` allows the user to install callbacks for when
a file or module changes, and also enters a blocking loop
that catches any errors that might arise from the callbacks.

This new `add_callback` function allows installing callbacks and keeping
them around in the background. This commit also reimplements
`entr` in terms of `add_callback` and `remove_callback`, and adds tests.