Insert pkgdata after initialization is complete
Formerly we inserted `pkgdatas[id] = pkgdata` when we first
created `pkgdata`, and then parsed the package.
This PR changes that to do the insertion at the end.
This allows one to wait on `haskey(pkgdatas, id)` and be
confident that the initial setup of the package is complete.
We use this in `add_require` to make sure we add
atsign-require blocks after the rest of the package is in place.

This is a step in the direction of fixing #431, but it's not finished yet.