Shorten startup time by defering compilation
This defers compilation of `revise` and `queue_includes!`.
Startup no longer has a delay before showing you the 2nd REPL
prompt, and the hit for the next package load is roughly a
third of a second. Formerly these were multiple seconds.

This code still needs to be compiled if you trigger a revision,
but the reduced overhead means that the penalty for using package
is low unless it's also providing added value.

Also, since `revise` no longer runs on each REPL input, the warning
is not repeated, and this requires changes to the tests.