Circumvent CI hang on 1.8 (#722)
#710 attempted to debug a Requires-based hang that seems
to only occur on CI with Julia 1.8. It may be due to lock contention,
but the precise case was not obvious.

This just disables the problematic test.
Since it doesn't happen on nightly, and since Requires will be
supplanted by the new weakdeps mechanism, we can direct any
users who encounter this bug to the alternatives.