Fix invalidation in `eval_rex` (#729)
This invalidation was triggered by
    convert(::Type{T}, i::LoopVectorization.UpperBoundedInteger) where T<:Number
It turns out `eval_rex` had a `Core.Box` and various issues stemming from
too many small Unions for inference to succeed. Some of the later calls
imply that the type must be specific (there are no methods that can
handle the alternative type). By type-asserting that type early we avoid
the inference failures.