Restore filtering of non-src files
This patch restores the filtering of non-src files that was lost
in #767. After https://github.com/JuliaLang/julia/pull/49866,
`Base.parse_cache_header` returns both the non-filtered and filtered
list instead of doing filtering internally based on the
`srcfiles_only::Bool` keyword argument.

There is a test for this failure mode that fails on Revise master branch
and passes with this patch (see
https://github.com/timholy/Revise.jl/blob/1059181bed06387e9fbcea137dce28a80c5c45d9/test/runtests.jl#L3070-L3080).