Reduced reassignment to global consts
When a single top-level expression contains both `const` definitions and methods that use those consts, the `add_named_dependencies!` step of `lines_required!` can induce a dependency on the statement assigning to the `const`. This is problematic when your only goal is to parse signatures, as it can trigger overwriting of `consts` that may be needed.

Fixes #789