Adjust to upcoming change in world age semantics
Upcoming base PR https://github.com/JuliaLang/julia/pull/56509 makes precise
when exactly world age increments automatically at top-level. Packages are
mostly not supposed to notice, but of course Revise is a bit patholgical in
that it does basically expect asynchronous changes to have top level effects.
After that PR, that requires an explicit opt-in. This PR makes the appropriate
adjustments to the tests.