Fix "juliadir" test
The test was ineffective because it was relying on binding replacement,
which is undefined behavior on all versions of Julia (ironically it will
not be in 1.12, but that change is not done yet). As a result,
the redefinition of `juliadir` was not taking effect and the
test was simply running with the old juliadir. Change `juliadir`
to a (typed) global and update the test to make sure it's actually
working.