update to JuliaInterpreter 0.10 (#911)
* update to JuliaInterpreter 0.10

This commit implements the migration to the new JuliaInterpreter
interface proposed in JuliaDebug/JuliaInterpreter.jl#683.

It purely performs the migration to the new interface and does not
include any refactoring based on it.

In practice, `methods_by_execution!` is quite complex, and using the
`Interpreter` interface may not allow us to simplify its implementation.

That said, this commit seems to achieve a modest latency improvement by
changing the argument type declaration from `@nospecialize(recurse)` to
`interp::Interpreter` (with easier code specialization).

* checking in docs/Manifest.toml