remove the interface for supporting non-jl files
Remove interface for supporting non-.jl files introduced in timholy/Revise.jl#680.
This interface forces functions that used to handle only `AbstractString`
to deal with arbitrary types, making refactoring across the Revise.jl
codebase difficult.
The simplified interface proposed in the previous commit is also
fragile since it relies on initialization and runtime-mutable data
structures, and so it is therefore not a complete solution.
After discussion, and since this experimental interface has effectively
no users, we remove it for now to enable codebase refactoring until
it becomes necessary again.