Track Core.Compiler under a separate PkgId
It shares a UUID with the Compiler stdlib (which depending on version
may or may not delegate to Core.Compiler). I think this should work
because in the non-delegation case, the Compiler stdlib doesn't actually
load any files and in the delegation files they are totally separate,
but let's keep an eye on any Revise weirdness around the Compilers.