Preliminary support for `include(mapexpr, file)`
In conjunction with `CodeTracking.MapFile`, this allows Revise to track
changes in files that are included with `include(mapexpr, file)`. This
PR supports the implementation of revisions for such files, properly
applying the `mapexpr` transformation to the included file's
expressions.

What's missing:
- Revise can't correctly populate the internal data structures needed to
  initiate `mapexpr` without help. The problem is that Revise only
  parses files that get edited, but discovering that `file2.jl` gets
  included from `file1.jl` via a `mapexpr` requires parsing `file1.jl`,
  which may not have been edited and therefore never scanned via Revise.
  The right way to fix this is in Julia itself, writing additional info
  to the package cache header. This will not land before Julia 1.13.
- `include(mapexpr, file)` is well-suited to include the same file more
  than once with different `mapexpr`s. While it hasn't been tested, this
  seems unlikely to work for now, and the anticipated problems stem all
  the way back to Julia itself (again, issues that should be fixed in
  the package cache header).

Package authors who want to use this in their packages can add a call to
`Revise.parseall(@__MODULE__)` in their `__init__` function to
ensure that the `mapexpr` files are parsed and tracked by Revise.

Fixes #820, #634