Fix intermittent Base CI failure due to misdetected juliadir
We have an intermittent Revise test failure in Julia base CI, e.g.
https://buildkite.com/organizations/julialang/pipelines/julia-master/builds/49637/jobs/0198728e-c6d1-4db8-8dd8-84eb87f72c6d/log
```
7-element Vector{Tuple{String, String}}:
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/./build_h.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/build_h.jl")
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/./version_git.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/version_git.jl")
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/./pcre_h.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/pcre_h.jl")
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/./errno_h.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/errno_h.jl")
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/./uv_constants.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/uv_constants.jl")
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/./file_constants.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/file_constants.jl")
 ("/cache/build/builder-amdci5-3/julialang/julia-master/base/features_h.jl", "/cache/build/builder-amdci5-3/julialang/julia-master/base/features_h.jl")
```

The apparent cause is that the Revise test job runs on the same
node that built the original binary. Upon further investigation,
that cause is apparent - when setting `juliadir`, we only check
whether the original build dir exists, not whether it actually
belongs to the julia executable we're testing. On the buildbots,
they will happily accept the next build job, causing the path
to exist again (in the same location), but actually we're not
testing a source build. Fix this by always trying to use the
canonical build directory (which should be created by the build
system), even in a source build. If it doesn't exist (e.g.
because the build systme ran incompletely), we try to fallback
to the original directory. If an only if `Sys.BINDIR` points
to where it would in an ordinary source build. Lastly we fall
back to the ascending search procedure, though I'm a bit
skeptical about it, since it could easily find the wrong
installation. I added a warning for users to let us know if
they encounter this frequently.