Rename `CodeTrackingMethodInfo` to `MethodInfo`
Consolidate the two `MethodInfo` types into one, as managing two
`MethodInfo` implementations is tedious, so let's make it so we can
manage only the full method info.

Previously there was a simple `IdDict`-based `MethodInfo` in lowered.jl
for testing and the full `CodeTrackingMethodInfo` in packagedef.jl. This
removes the simpler version and renames `CodeTrackingMethodInfo` to
`MethodInfo`.